added some types, broken NameForm out into a separate file
