+++ components/gcp-click-to-deploy/src/App.tsx
@@ -1,3 +1,4 @@
+import * as jsYaml from 'js-yaml';
 import * as React from 'react';
 
 // TODO(jlewi): Replace with official logo image. Also it would 
@@ -13,9 +14,9 @@ import logo from './logo.jpg';
 import clusterSpecPath from './configs/cluster-kubeflow.yaml';
 import clusterJinjaPath from './configs/cluster.jinja';
 
-import * as jsYaml from 'js-yaml';
-
 import './App.css';
+
+import NameForm from './NameForm';
 
 // TODO(jlewi): Can we set email automatically to the signed in email?
 // TODO(jlewi): For the FQDN we should have a drop down box to select custom domain or automatically provisioned
@@ -23,7 +24,6 @@ import './App.css';
 // <hostname>.endpoints.<Project>.cloud.goog
 
 let gapi: any = null;
-
 let auth2: any = null;
 
 // TODO(jlewi): ClientId for project cloud-ml-dev we should change this.
@@ -43,195 +43,9 @@ const updateSignIn = () => {
   log('update sign in state');
   if (auth2.isSignedIn.get()) {
     log('signed in');
-    // var emailObj = window.document.getElementById('email');
-
-    // if (emailObj.value == '') {
-    //   console.log('Setting email');
-    //   emailObj.value = 'newvalue@.com';
-    // }
-
-    // TODO(jlewi): Code copied from the google+ example
-    //  helper.onSignInCallback(gapi.auth2.getAuthInstance());
   } else {
     log('signed out');
-    // TODO(jlewi): Code copied from the google+ example
-    // helper.onSignInCallback(gapi.auth2.getAuthInstance());
   }
-}
-
-class NameForm extends React.Component {
-  public handleChange: any;
-  public state: any;
-  public props: any;
-
-  constructor(props: any) {
-    super(props);
-    this.state = {
-      deploymentName: 'kubeflow',
-      email: 'john@doe.com',
-      hostname: '<HOST>.endpoints.<PROJECT>.cloud.goog',
-      ipName: 'kubeflow',
-      project: 'cloud-ml-dev',
-      zone: 'us-east1-d',
-    };
-
-    // startApp();
-
-    this.handleChange = this._handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  // Create a  Kubeflow deployment.
-  public createDeployment() {
-    const templates = this.props.getDeployemtTemplates();
-    if (this.state.project === '') {
-      alert('You must set project');
-      return
-    }
-
-    if (this.state.zone === '') {
-      alert('You must set zone');
-      return
-    }
-    if (this.state.email === '') {
-      alert('You must set email');
-      return
-    }
-    if (this.state.ipName === '') {
-      alert('You must set IP Name');
-      return
-    }
-    if (this.state.deploymentName === '') {
-      alert('You must set deployment name');
-      return
-    }
-    if (this.state.hostName === '') {
-      alert('You must set hostname');
-      return
-    }
-
-    const kubeflow = templates.clusterSpec.resources[0];
-
-    kubeflow.name = this.state.deploymentName;
-    kubeflow.properties.zone = this.state.zone;
-
-    // Load the bootstrapper config.
-    const config: any = jsYaml.safeLoad(kubeflow.properties.bootstrapperConfig);
-
-    if (config == null) {
-      alert('Property bootstrapperConfig not found in deployment config.');
-      return;
-    }
-
-    const state = this.state;
-    config.app.parameters.forEach((p: any) => {
-      if (p.name === 'acmeEmail') {
-        p.value = state.email;
-      }
-
-      if (p.name === 'ipName') {
-        p.value = state.ipName;
-      }
-
-      if (p.hostname === 'hostname') {
-        p.value = state.hostname;
-      }
-    });
-
-    templates.clusterSpec.resources[0] = kubeflow;
-    const yamlClusterSpec = jsYaml.dump(templates.clusterSpec);
-    this.props.appendLine('Spec:\n' + jsYaml.dump(yamlClusterSpec));
-
-    const resource = {
-      'name': this.state.deploymentName,
-      'target': {
-        'config': {
-          'content': yamlClusterSpec,
-        },
-        'imports': [
-          {
-            'content': templates.clusterJinja,
-            'name': 'cluster.jinja',
-          }
-        ],
-      },
-    };
-
-    const appendLine = this.props.appendLine;
-    gapi.client.deploymentmanager.deployments.insert({
-      project: this.state.project,
-      resource,
-    }).then((res: any) => {
-      appendLine('Result of insert:\n' + res);
-    }, (err: any) => {
-      appendLine('Error doing insert:\n' + err);
-      alert('Error doing insert: ' + err)
-    });
-
-  } // insertDeployment
-
-  public render() {
-    return (
-      <div>
-        <form onSubmit={this.handleSubmit}>
-          <div>
-            <label>
-              Project:
-          <input type='text' name='project' value={this.state.project} onChange={this.handleChange} />
-            </label>
-          </div>
-          <div>
-            <label>
-              Deployment name:
-          <input type='text' name='deploymentName' value={this.state.deploymentName} onChange={this.handleChange} />
-            </label>
-          </div>
-          <div>
-            <label>
-              Zone
-          <input type='text' name='zone' value={this.state.zone} onChange={this.handleChange} />
-            </label>
-          </div>
-          <div>
-            <label>
-              IP name:
-          <input type='text' name='ipName' value={this.state.ipName} onChange={this.handleChange} />
-            </label>
-          </div>
-          <div>
-            <label>
-              hostname:
-          <input type='text' name='hostname' value={this.state.hostname} onChange={this.handleChange} />
-            </label>
-          </div>
-          <div>
-            <label>
-              Email for Lets Encrypt:
-          <input type='text' name='email' value={this.state.email} onChange={this.handleChange} />
-            </label>
-          </div>
-        </form>
-        <div>
-          {/* TODO: REVISIT */}
-          <button className='createDeployment' onClick={this.createDeployment}>
-            Create Deployment
-      </button>
-        </div>
-      </div>
-    );
-  }
-  private _handleChange(event: Event) {
-    const target = event.target as any;
-    this.setState({
-      [target.name]: target.value
-    });
-  }
-
-  // TODO(jlewi): We aren't really using submit since we don't want to post the data anywhere.
-  private handleSubmit(event: any) {
-    event.preventDefault();
-  }
-
 }
 
 class SignInButton extends React.Component {
@@ -247,8 +61,13 @@ class SignInButton extends React.Component {
   }
 }
 
-class App extends React.Component {
-  public state: any;
+interface AppState {
+  clusterJinja: string;
+  clusterSpec: any;
+  gapiLoaded: boolean;
+}
+
+class App extends React.Component<any, AppState> {
   constructor(props: any) {
     super(props);
     this.state = {
@@ -258,7 +77,7 @@ class App extends React.Component {
     };
   }
 
-  public getDeployemtTemplates() {
+  public getDeploymentTemplates() {
     return {
       clusterJinja: this.state.clusterJinja,
       clusterSpec: this.state.clusterSpec,
@@ -275,7 +94,6 @@ class App extends React.Component {
     gapi.load('auth2' /* Initialize the API client library */, null);
 
     //   //On load, called to load the auth2 library and API client library.
-    const thisVar = this;
     gapi.load('signin2', () => {
       log('Loaded signin2');
     });
@@ -287,8 +105,6 @@ class App extends React.Component {
         log('Loaded deploymentmanager');
         // If we don't set a var = to this then we can't bind this in the
         // callback.
-        thisVar.state.clientLoaded = true;
-
 
         log('load auth2');
         // See: https://developers.google.com/identity/sign-in/web/reference
@@ -335,7 +151,9 @@ class App extends React.Component {
     if (script.getAttribute('gapi_processed')) {
       gapi = (window as any).gapi;
       log('Client is ready! Now you can access gapi. :)');
-      this.state.gapiLoaded = true;
+      this.setState({
+        gapiLoaded: true,
+      });
       this.createClients();
       if (window.location.hostname === 'localhost') {
         gapi.client.load('http://localhost:8080/_ah/api/discovery/v1/apis/metafields/v1/rest')
@@ -378,7 +196,6 @@ class App extends React.Component {
     // TODO(jlewi): The fetches should happen asynchronously.
     // The user shouldn't be able to click submit until
     // the fetches have succeeded. How can we do that?
-    const n = this;
 
     this.appendLine('loadClusterJinjaPath');
     // Load the jinja template into a string because 
@@ -389,7 +206,9 @@ class App extends React.Component {
         return response.text();
       })
       .then((text) => {
-        n.state.clusterJinja = text;
+        this.setState({
+          clusterJinja: text,
+        });
         log('Loaded clusterJinja successfully');
       })
       .catch((error) => {
@@ -405,7 +224,9 @@ class App extends React.Component {
         return response.text();
       })
       .then((text) => {
-        n.state.clusterSpec = jsYaml.safeLoad(text);
+        this.setState({
+          clusterSpec: jsYaml.safeLoad(text),
+        });
         // log('Loaded clusterSpecPath successfully');
       })
       .catch((error) => {
@@ -442,7 +263,7 @@ class App extends React.Component {
             * Click create deployment
         </p>
         {/* tslint:disable-next-line:jsx-no-lambda */}
-        <NameForm id='nameForm' appendLine={(newline: any) => this.appendLine(newline)} getDeployemtTemplates={() => this.getDeployemtTemplates()} />
+        <NameForm className='nameForm' appendLine={(newline: any) => this.appendLine(newline)} getDeploymentTemplates={() => this.getDeploymentTemplates()} />
         <p> Logs </p>
         <textarea id='logs' data-width='800' data-height='800' />
       </div>
