+++ components/gcp-click-to-deploy/tsconfig.json
@@ -16,11 +16,7 @@
     "noImplicitAny": true,
     "strictNullChecks": true,
     "suppressImplicitAnyIndexErrors": true,
-    "noUnusedLocals": true,
-    "types": [
-      "gapi",
-      "gapi.auth2"
-    ]
+    "noUnusedLocals": true
   },
   "exclude": [
     "node_modules",
